llm-based-assistants/sessions/block2_core/2611_2711_0312_0412/requirements.txt
ddgs
langchain_community
langchain_nvidia_ai_endpoints
langgraph
python-dotenv
pytidylib